requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
fsspec==2021.5.0
future==0.18.2
google-auth==1.30.1
google-auth-oauthlib==0.4.4
grpcio==1.38.0
idna==2.10
ipykernel==5.5.5
ipython==7.24.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.20.3
oauthlib==3.1.1
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.18
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyDeprecate==0.3.0
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytorch-lightning==1.3.3
pytz==2021.1
pywin32==301
pywinpty==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
Send2Trash==1.5.0
six==1.16.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
terminado==0.10.0
testpath==0.5.0
threadpoolctl==2.1.0
torch==1.8.1
torchmetrics==0.3.2
torchvision==0.9.1
tornado==6.1
tqdm==4.61.0
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
yarl==1.6.3
